Выделяет интерфейсы в отдельный файл для каждого
