+++ FiveWords.Overall/Repository/Interfaces/IOnePasswordRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace FiveWords.Repository.Interfaces;
+
+public interface IOnePasswordRepository : IBaseRepository
+{
+    byte[] GetPasswordHash();
+    void SavePasswordHash(byte[] data);
+}
